fix: Link in pool detail
